requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.3
django-environ==0.11.2
future==1.0.0
gunicorn==23.0.0
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
numpy==2.2.0
packaging==24.2
pillow==10.4.0
proglog==0.1.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.8.2
